make linter easy to include into other software
Point of this change is to make the linter to be easy to import other golang
projects. Situation in Cloudflare matches well with this scenario. A
domain-connect server software should not rely on external executable when
checking validity of template input. It is better if template input can be
validated straight from input stream provided to the server software.

This change also includes few golangci-lint fixes.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>