check validity of template variables
The template variables need to be strings with a valid character set starting
from '%' and terminated by '%'.  Pawel proposed valid characters to be [a-zA-Z]
but templates also use [0-9] and '-' (hyphen) and '_' (underscore) in some cases.
It is probably too late to ban numbers and couple delimiters.

Proposed-by: Pawel Kowalik
Reference: https://github.com/Domain-Connect/dc-template-linter/issues/7
Signed-off-by: Sami Kerola <kerolasa@iki.fi>