add groupId to record message printout
When a template has lots of record entries there tends to be groupId in use
to tell which ones to apply.  These groupId fields are pretty nice locator
when trying to locate an entry that is raising a warning or error.

In case groupId is not defined and message is submitted field will log blank.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>