allow ttl fixing with new command-line option
Include support to add ttl values to records that are missing one or the
value is invalid.  This change also fixes inplace support that earlier added
zero ttl to records that do not require one, such as REDIR30x and SPFM
types.  Existing unncessary TTLs will be left in place.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>