spec: allow number types to be ints, strings, and/or variables
After this change numbers are always quoted after reformatting.  That is the
only way to be somewhat consistent what is the representation format when
%variables% need to be supported, and they need to be strings.

As a practical example, one could have the following in a template (and it
would fail, but it's obvious why - see the hostname).

    "records": [
        {
            "type": "A",
            "host": "the-ttl-is-plain-number-this-is-ok",
            "ttl": 1800
        },
        {
            "type": "A",
            "host": "the-ttl-is-quoted-number-this-is-ok",
            "ttl": "8100"
        },
        {
            "type": "A",
            "host": "the-ttl-is-variable-this-is-ok",
            "ttl": "%varttl%"
        },
        {
            "type": "A",
            "host": "ttl-is-BROKEN-it-cannot-be-just-a-string",
            "ttl": "just-a-string"
        }
    ]

Reference: https://github.com/Domain-Connect/spec/pull/100
Signed-off-by: Sami Kerola <kerolasa@iki.fi>